combined UI with check-in endpoint
