--- src/app/check-in/page.tsx
@@ -1,9 +1,0 @@
-import CheckInPage from "./CheckInPage";
-
-export default function CheckIn() {
-  return (
-    <main className="w-full bg-pastel-pink">
-      <CheckInPage />
-    </main>
-  );
-}
